Inital Entro-Refactor Commit
